requirements.txt
absl-py==2.0.0
alembic==1.12.1
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
category-encoders==2.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.7.0
comm==0.1.4
contourpy==1.2.0
cubinlinker-cu11==0.3.0.post1
cuda-python==11.8.2
cudf-cu11==23.10.0
cuml-cu11==23.10.0
cupy-cuda11x==12.2.0
cycler==0.12.1
Cython==3.0.5
dask==2023.9.2
dask-cuda==23.10.0
dask-cudf-cu11==23.10.0
debugpy==1.8.0
decorator==5.1.1
distributed==2023.9.2
exceptiongroup==1.1.3
executing==2.0.1
fastrlock==0.8.2
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.0
fsspec==2023.10.0
gast==0.5.4
gensim==4.3.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.3
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.5.0
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
lightgbm==4.1.0
llvmlite==0.40.1
locket==1.0.0
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mol2vec @ git+https://github.com/samoturk/mol2vec@850d944d5f48a58e26ed0264332b5741f72555aa
msgpack==1.0.7
nest-asyncio==1.5.8
numba==0.57.1
numpy==1.24.4
nvtx==0.2.8
oauthlib==3.2.2
opt-einsum==3.3.0
optuna==3.4.0
packaging==23.2
pandas==1.5.3
parso==0.8.3
partd==1.4.1
patsy==0.5.3
pexpect==4.8.0
Pillow==9.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.6
ptxcompiler-cu11==0.7.0.post1
ptyprocess==0.7.0
pure-eval==0.2.2
py-boost==0.4.3
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.16.1
pylibraft-cu11==23.10.0
pynvml==11.4.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
raft-dask-cu11==23.10.0
rdkit==2023.9.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rmm-cu11==23.10.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
stack-data==0.6.3
statsmodels==0.14.0
tblib==3.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.14.1
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
treelite==3.9.1
treelite-runtime==3.9.1
typing_extensions==4.8.0
ucx-py-cu11==0.34.0
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.14.1
zict==3.0.0
zipp==3.17.0
